+++ app/src/main/java/protect/card_locker/DataFormat.java
@@ -2,7 +2,8 @@ package protect.card_locker;
 
 public enum DataFormat
 {
-    CSV,
-
+    Catima,
+    Fidme,
+    VoucherVault
     ;
 }