+++ app/src/main/java/protect/card_locker/ImportExportTask.java
@@ -13,6 +13,9 @@ import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.nio.charset.Charset;
 
+import protect.card_locker.importexport.MultiFormatExporter;
+import protect.card_locker.importexport.MultiFormatImporter;
+
 class ImportExportTask extends AsyncTask<Void, Void, Boolean>
 {
     private static final String TAG = "Catima";
@@ -58,16 +61,8 @@ class ImportExportTask extends AsyncTask<Void, Void, Boolean>
     {
         boolean result = false;
 
-        try
-        {
-            InputStreamReader reader = new InputStreamReader(stream, Charset.forName("UTF-8"));
-            result = MultiFormatImporter.importData(db, reader, format);
-            reader.close();
-        }
-        catch(IOException e)
-        {
-            Log.e(TAG, "Unable to import file", e);
-        }
+
+        result = MultiFormatImporter.importData(db, stream, format);
 
         Log.i(TAG, "Import result: " + result);