+++ app/src/main/java/protect/card_locker/preferences/Settings.java
@@ -90,4 +90,14 @@ public class Settings
     {
         return getBoolean(R.string.settings_key_lock_barcode_orientation, false);
     }
+
+    public boolean getKeepScreenOn()
+    {
+        return getBoolean(R.string.settings_key_keep_screen_on, true);
+    }
+
+    public boolean getDisableLockscreenWhileViewingCard()
+    {
+        return getBoolean(R.string.settings_key_disable_lockscreen_while_viewing_card, true);
+    }
 }